quic: inline and const where possible
